refactor: 재사용 위해 useTerm 분리
